+++ src/sys/unix/mod.rs
@@ -119,11 +119,22 @@ impl SerialPort {
 	}
 
 	pub fn set_configuration(&mut self, settings: &Settings) -> std::io::Result<()> {
+		let initial_settings = settings;
 		let mut settings = settings.clone();
+
+		// Ensure input and output processing is disabled.
+		// TODO: Should we just expose "make_raw()" and push this burden to the user?
 		unsafe {
 			libc::cfmakeraw(&mut settings.termios as *mut _ as *mut libc::termios);
 		}
+		settings.set_baud_rate(initial_settings.get_baud_rate()?)?;
+		settings.set_char_size(initial_settings.get_char_size()?);
+		settings.set_stop_bits(initial_settings.get_stop_bits()?);
+		settings.set_parity(initial_settings.get_parity()?);
+		settings.set_flow_control(initial_settings.get_flow_control()?);
+
 		settings.set_on_file(&mut self.file)?;
+
 		let applied_settings = self.get_configuration()?;
 		if applied_settings != settings {
 			Err(other_error("failed to apply some or all settings"))